feat：新增Flink API模板类；重构DimApp代码；修复维度数据预加载bug；新增JDBC工具类
