+++ realtime-common/src/main/java/com/liangheee/gmall/realtime/common/constant/Constant.java
@@ -5,11 +5,12 @@ package com.liangheee.gmall.realtime.common.constant;
  * * @date 2024/10/25
  */
 public class Constant {
+    public static final String HADOOP_USER_NAME = "liangheee";
     public static final String TOPIC_DB = "topic_db";
     public static final String BROKER_SERVERS = "hadoop102:9092,hadoop103:9092,hadoop104:9092";
     public static final String MYSQL_HOST = "hadoop102";
     public static final int MYSQL_PORT = 3306;
-    public static final String MYSQL_USERNAME = "root";
+    public static final String MYSQL_USER = "root";
     public static final String MYSQL_PASSWD = "123456";
     public static final String MYSQL_URL = "jdbc:mysql://hadoop102:3306?useSSL=false";
     public static final String HBASE_HOST = "hadoop102,hadoop103,hadoop104";
