--- realtime-common/src/main/java/com/liangheee/gmall/realtime/common/utils/JDBCUtil.java
@@ -1,9 +1,0 @@
-package com.liangheee.gmall.realtime.common.utils;
-
-/**
- * @author liangheee
- * * @date 2024/10/28
- */
-public class JDBCUtil {
-
-}
